Test is now functioning though what I want to be injected is not. GeoCamMemo is being injected w/ DjangoMemoImplementation and not FakeDjangoMemoImplementation. Unsure as to what's causing this. Need to more closely examine documentation and examples at http://pivotal.github.com/robolectric/roboguice.html.
The issue Anooj and I were having last night seems to be caused by our lack of understanding on how Robolectric is creating "shadow" objects. Shadow objects in Robolectric are standins for typical Android objects. Not sure if this is why we needed to be more explicit in our listView.setAdapter or if what we had would have died on a normal run as well.

I also upgrade to the Robolectric RC1 jar. This could have also fixed the issue as I did the upgrade as I integrated the new code to create items for the ListView.